imm: ignore searching for system attributes in class definition [#2897]
IMM fails to load model file which has system attribute(s) in object tags
but not in its class element. In case of importing, it always
fails to import the file having system attributes in object tags.

This patch ignores checking about whether the attribute name exists in
class element if the processing attribute name in object tags is one of
system attributes.