rded: perform KV store operations outside main thread [#2905]
Sometimes the 'watch' command in the KV plugin will not return
a takeover request, if the KV store does not respond in time.
Then rded would try to read the takeover request in the main
thread after receiving a takeover request notification.

This can cause rded to not respond to AMF callbacks in a timely
fashion. We should try to read the takeover request in another
thread to avoid this.