log: fix logtest 4 58 fail [#2909]
Due to hard-disk problem, the log file has not been closed successful and
log file has not been renamed. When previous test case (4 48) closes log file fail,
test case (4 58)  also count the size of (4 48)'s log file. This cause the size of
log file is not 0.

The patch update to check (grep) if the log record is in log file or not instead
count the log file size.