mbc: fix mbcsv loop forever while it is being dispatch ALL [#2899]
When processing "MBCSV_PEER_UP_MSG" msg in case dispatch all from
user, the msg may be too old. The msg may be come from the node
that already rebooted. This reason cause mbcsv loop forever in WHILE
loop, because the active node cannot find the adest of peer msg.

The solution is that find entity in peer list and compare if the peer
node id that already exist. If the peer node id exist in the entity
(get node id from peer anchor), this mean that mbcsv already processed
the "MBCSV_PEER_UP_MSG" for that peer node.