log: fix timeout when deleting configuration stream [#2903]
During closing stream, log sevice does try again in 8s when renaming file.
the log server is hung in that time and cannot poll and process any requests.
If the completed_callback come from imm in that time, the callback cannot be
processed in time (in 6s) and cause timeout issue happen.

The try again in 8s should be remove to avoid hanging any requests. And specialy
here is to avoid timeout of callback requests from imm