amfd: Set SA_AMF_READINESS_IN_SERVICE for qualified SU after cluster startup timeout [#2916]
In the scenario of single step upgrade where the UNLOCK-IN/UNLOCK
admin op are issued to a SU hosted on non-active node while cluster
startup timer is active and not all ncs SU on that node are fully
assigned. In such case, amfd currently accepts the UNLOCK admin op,
change AdminState to UNLOCKED but the ReadinessState is still OUT_OF_SERVICE.
When the cluster startup timeout, amfd is not giving assignment to
this SU which is still OUT_OF_SERVICE. amfd should return TRY_AGAIN
in the first place of UNLOCK command since the node is still DISABLED
and ABSENT since all ncs SUs are not assigned, but it would cause
the upgrade delays.

When cluster startup timer expires, the patch checks if SU is qualified
to be IN_SERVICE but its readiness state is still OUT_OF_SERVICE, set
it to be IN_SERVICE before amfd starts assignments.