plmd: fix crash when saPlmReadinessTrack is called in error [#2919]
plmd crashes when saPlmReadinessTrack is called with entities pointer set,
but smaller than what plmd would return.

In this case plmd is returning ERR_NO_SPACE, which is correct, but it is
setting numberOfEntities without setting the entities pointer. This causes
the edu routines to crash.

It is not necessary to set numberOfEntities since we are returning an error
code.