amfd: reboot nodes that report conflicting 2N active assignments [#2920]
After a split network event, both SCs can reboot endlessly,
due to this assertion:

2018-08-29 18:05:34.689 SC-2 osafamfd[263]: src/amf/amfd/sg_2n_fsm.cc:596:
  avd_sg_2n_act_susi: Assertion 'a_susi_1->su == a_susi_2->su' failed.
2018-08-29 18:05:34.695 SC-2 osafamfnd[273]: ER AMFD has unexpectedly crashed. Rebooting node

During the network split, a SC could assign another SU to be active,
if the node hosting the old active 2N assignment is not reachable.

The assert occurs after the network is merged. SC absence must be
enabled.

For now, we can aid recovery of the cluster by rebooting
both of the PLs in place of the assertion.