plmd: fix adding and removing of invocation id to list [#197]
Jan 22 11:09:03 localhost osafplmd[3988]: Invocation id mentioned in the resp, is not found in the grp->inocation_list. inv_id: 9

If multiple entities are part of the same entity group, and START or VALIDATE
tracking is requested, if an admin operation is done on these entities, once
one response is sent the other responses are ignored. But, the entities that
didn't return a successful response all report "Admin operation can not be
performed" because they failed to process the tracking response. This is
because when the first invocation id is removed from the list, all the others
are removed, too. Now those entities are stuck in this bad state.

Fix the remove routines so that only the invocation in the response is
removed from the list.