ntf: Refactor NtfReader filter to fix incorrect searchcriteria [#2925]
NtfReader filters with searchcriteria NtfAfterTime, then creates a list of cached notification.
But the filter() is not correct.

The patch simplifies the unncessary complexity of filter classes by replacing 'switch' statements
inside filterSearchCriteria(). The bug reported in this ticket (due to NtfAfterTime::filter) is
also fixed along with the new refactoring code.