imm: fix osafimmnd coredump genereted during sanity test [#2947]
The coredump is generated in the context of processing the message type
"IMMND_EVT_D2ND_IMPLDELETE" because the memory is corrupted at the time
of decoding that message.

It allocated 'size' bytes of memory with the boundary in range [0 - 'size - 1'],
but modified - added null terminated, the memory at the index of `size` which
was out of that range.

This patch fixes such issue. The memory should be allocated with `size + 1`
bytes in length.