amfd: add class definitions for new timers [#2918]
osafAmfDelayNodeFailoverTimeout - the number of seconds we wait
after MDS down is received before we consider it truly down.

osafAmfDelayNodeFailoverNodeUpWait - the number of seconds we
wait for Node Up after receving MDS up, before we send reboot
to the node. After sending  reboot to a node, also wait up to
this number of seconds before we consider the node to be
down (unless MDs down is received first).