imm: fix introduceMsg resent in every 0.1 second [#2959]
It is expected that resending introduceMsg should happen every 05 seconds
if the previous send has not yet received the response e.g because of having
network issue or having the active IMMD unresponsive. However, in some cases,
the re-send is done every 0.1 second.

This patch adds a few code to make the logic work as expectation.