amfd: Give assignment for pre-instantiated su after the node joins cluster [#2960]
If the SU is unlock-in/unlock before the node joins cluster, the SU is instantiated
and in unlocked state. However, when the node completes joining the cluster, amfd
assumes all applications' SU uninstantiated and starts the instantiation, thus the
instantiated/unlocked SU is forgot to give the assignments.

The patch handles the above case by giving assignments to the instantiated/unlocked
SU.