osaf: ensure takeover_requests have a lease [#2954]
In CreateTakeoverRequest(), if the initial attempt fails,
then the takeover_request is created without a lease.

Furthermore, when the takeover_request result is set,
it is being set without a lease, and the takeover_request
is not automatically removed.

Add <timeout> parameter to KeyValue::Set, and remove
default value for the <timeout> parameter to KeyValue::Create to
ensure a timeout is always set.