imm: correct mismatched log info showing resources in displayverbose [#2986]
According to IMM documentation, it states that when the number of implementer
or admin owner resources is greater than 127, the verbose output is displayed
to syslog. However, the syslog messages contain incorrected information:
it says the number is 128 rather than 127 and implementer name instead of
admin owner name in case the resource display is about admin owner.

This patch corrects such mistakes.