dtm: Fix dtm close socket due to duplication of adding node IP info [#2984]
During cluster start, one node (node 1) broadcast up msg to other node. The
remote node (node 2) get this msg and send the connection to node 1 (connect()).
Similarly node 1 send the connection to  node 2 after node 2 broadcast up msg to.
Beside of node 2 connect() to node 1, node 2 also add the IP and ID info of node 1 to database.
But before of that, node 2 may also accept the connection that come from node 1. The
acception is also add node ID of node 1. So there is 2 times adding the node ID
info of node 1 to database in node 2. This causes the socket connection is closed
and node is  restart again.

The patch change to retrieve node from database by node IP instead node ID in
processing connection. This will reject the double of establishing connection
between 2 nodes and also double of adding node IP to database.