ntf: Limit the logger buffer [#2961]
When writing the notificaion fail with TRY_AGAIN in callback, the notificaion is pushed
again to the list. If this happens for long time, the list is going to be very big.
This cause NTFD take time to process writing all the notification in the list and
the request from NTFA come this time may be timeout.

The patch does:
   - Limit the logger buffer
   - Provide the env variable that user can set the value of the limit
   - Return TRY_AGAIN error in case the limit of buffer is reached and write all the
     notifications in the buffer to the log file. The current of notification isn't
     written to log file.