clmd: send PLM response if node down detected [#2998]
PLM HE/EE lock never transitions to LOCKED state if node crashes on which the
PLM lock is being done.

If a LOCK admin command from PLM is issued on a HE or EE, CLM sends a tracking
request to AMF to validate and start the shutdown. When the START track is sent
CLM waits for a response from AMF. Because the request is coming from PLM, the
time supervision is set to UNKNOWN, which means the response has no time limit.
If the node crashes before AMF responds to CLM, the PLM never gets a response
from CLM and doesn't transition the HE/EE to locked state.

CLM needs to send a response to PLM if it detects node down and there is a
pending PLM operation on that node.