rded: process takeover request without delay [#3005]
Currently, a takeover request is not sent to the main thread immediately
so that MDS messages related to topology changes are processed first.

If the plugin informs us it has lost connectivity to the consensus service
by returning 'UNDEFINED', or we prioritise the current active SC,
then we should not delay the request. This will speed up self-fencing
of the current active node (if required).