ntf: Fix to NTFD doesn't send the response if MDS_DEST of client is disconnected [#2973]
When the sending notification request come to NTFD , they are put in the mbx
to wait for being processed. if client of these request is down (MDS_DEST of this client
is disconnected), MDS thread receive the event down and also put to mbx. The priority
of event down and send request are the same. So NTF will process the send request
before the event down.

The patch fix to NTFD will not send the response of request if the MDS_DEST of client has
already disconnected.