nid: fix opensafd crashed during start-up [#3013]
There is a dependency b/w svc_monitor_thread and spawn_services.
The coredump happens when spawn_services is executed while
the thread has not yet started. In this case, data is sent to the
pipe but no one consumed it. When it comes to consume the data,
will get unexpected data and crash the program.

This patch ensures the things will happen in the right order:
svc_monitor_thread must be in ready state before spawn_services()
is executed.