imm: fix racing in sending discard-node during network split [#3012]
At the time of spliting the cluster into 02 partitions but keeping a node
such as PL-3 connecting with both partitions, just IMMND on PL-3 will get
discard-node messages from both active IMMD on partition #1 and from standby
IMMD on partition #2.

That race later on caused IMMND on PL-3 crashed due to the mismatch
found at finalize-sync.

This patch makes a minor change at standby IMMD - rather then sending the
discard-node message even in standby role, will put the message in queue
and only broadcast it when the standby is assigned to active.