osaf: improve response time in etcd3.plugin [#3016]
if the initial call to watch takeover request in etcd3.plugin
is made when etcd has already been shutdown (for example,
when etcd is running locally and the node is being shutdown),
the plugin should return 0 with a fake takeover request to ensure
rded shuts down promptly. Otherwise, it will keep calling
watch, delaying node shutdown.