log: fix logd crash due to well known stream has numOpeners=0 [#3018]
If client closes stream is successful in active node but it cannot be checkpointed
to standby node. This causes the "numOpeners" is different between active and standby.
It also different number of streams that client opened  between standby and active.
When the client open again that stream, the "numOpeners" is updated to be equal between
active and standby while the client own number of streams is still different betwwen
active and standby. This will causes the logd crash.

Should restart standby node if handling the checkpoint of reopening stream
with different of "numOpeners" between active and standby node.