imm: fix coredump on immnd during upgrade [#3030]
The assertion about the existence of the new added attribute
`saImmFileSystemStatus` should be done if the admOp ID is
either 400 or 401.

Also correct the value of OPENSAF_IMM_FLAG_PRT51906_ALLOW
which presents the bitmask index for the new protocol.
The index should be 11, and therefore the value is 0x00000400.