amfnd: don't attempt su failover if active controller is rebooting [#3035]
In N+M model CSI-remove responses can get lost if active controller reboots.
In this case SG will be stuck in unstable state, and standby will never get
assignments.

We are the active controller, active for N+M, SU failover is set, and
failfast on termination failure is set for the nodes. If a component in the
SU crashes, and another component fails during cleanup, the node does
failfast. It currently attempts to do su failover in this case, but the
csi-remove responses from the payload can get lost because we are rebooting.
They eventually show up on the new active, but we get message-id errors.

Set a flag when the active controller is about to reboot. If the flag is set,
then don't do SU failover. Let the new active take care of the failover.