msgd: narrow ntf filter [#3034]
Can see the following log during controller switchover with lots of AMF
objects in the config: saNtfDispatch failed: 6.

msgd is listening for every NTF notification which occurs in the system. If
there is a failover there can be a large number of notifications which get
generated.

msgd should setup the filter when it registers with NTF to only get
notifications for the ones it cares about.