amf: remove SUSI assignemnt of dependent SI during failover [#3049]
When lock node invokes on active assignment. The dependent SI
follow with sponsor SI move to QUIESCED. There is a case that
the active assignment for sponsor is happening on remain SC node.
And that remaining node was down. The remove SISU only happen for
sponsor SI.
The fix is to remove SUSI of dependent SI.