amfnd: reset transition descriptor during comp restart [#3103]
If a component is configured to restart, instead of failover, on failure,
the previous transition descriptor is passed to the CSI set callback after
the restart.

The transition descriptor is not reset by amfnd in this case.

Always reset the transition descriptor to NEW_ASSIGN during a reassignment
due to restart.