log: make facility ID configurable [#3131]
Streaming log record is packaged in RFC5424 format and they all carry a fixed
facility ID (16). Therefore, log record receiver i.e. rsyslogd is not able to
filter log records on their facility id such as which ones are regular logs
or which ones are security.

The ticket is to have the facility ID configurable so that the receiver can
differentiate among types of log records based on that number.
Configure facility id via a new attribute `saLogStreamFacilityId` in class
`SaLogStreamConfig`. The valid range is [0-23]. The default value is 16 to keep
the streaming feature backward compatible. Always keep the value of attribute
`saLogStreamFacilityId` in sync with the standby. If user deletes value of the
attribute `saLogStreamFacilityId`, it will go back to the default value (16).