lck: fix infinite loop in lcknd when it cannot talk to lckd [#2813]
lcknd fails health check and restarts

When lcknd sends resource open requests to lckd it keeps them on a
list. If the open times out in lcknd because it cannot talk to lckd,
and the client finalizes or exits, lcknd attempts to clean up these
outstanding requests. But, it gets into an infinite loop traversing
the outstanding request list.

Fully traverse the list.