rded: Revert patch of #3142
The commit 8b3d5c24b16db4f0c6ac716dfd45b4e576d36e78 in patch #3142
helps the node to recover when the node finds another active.
However, it only helps when the plugin returns a value that caused
rded failed to takeover with return code 7 (INVALID_PARAM):

2020-01-16 10:18:31.103 SC-1 osafrded[183]: WA Takeover request failed (7)
2020-01-16 10:18:31.104 SC-1 osafrded[183]: WA Another controller is already active

In some other cases, rded fails to takeover with return code 14(EXISTED)
2020-02-07 14:03:59.208 SC-1 osafrded[159]: WA Takeover request failed (14)
2020-02-07 14:03:59.208 SC-1 osafrded[159]: WA Another controller is already active
And as legacy, the node was able to continue as standby role, but
now reboot due to #3142.