amfd: fix calculating standby rank for SIrankedSU with non-unique rank [#3149]
Standby rank which is passed to CSI set and protection group callbacks may not
be accurate.

If SIrankedSUs exist with non-unique ranks, AVD_SI::get_sisu_rank() is not
traversing all the SUs at that rank to determine the standby rank.

AVD_SI::get_sisu_rank() needs to traverse all the SUs at the particular rank.