amfnd: correct checking su assignement pending flag [#3176]
Su-si assigment flag only is set when amfnd receives request from amfd.
In SU restart escalation, the re-assignment is handled by amfnd internally.
And this flag is not set in this situation. When SU is in assigning after
SU restart due to escalation. The component failed and amfnd escalate it
to component failover. The Amfnd will try to mark su-si as assigned
temporaryly to remove assignment later. But amfnd crashes due to fail to
check su-si assignment flag.

In su-si assignment/removal response to AvD, the pending flag is only
checked if su_oper_state is enable.