ntf: set operation invoke name to unknown if failed to get it [#3178]
If ntfimcnd is restarted during ccb modify, the modify callback
will not contain the invoke name( implementer name or admin owner name).
In this case, the invoke name will be set to "unknown" and the ntfimcnd
can continue with sending notification.

The notification will contain only op that receive by the ntfimcnd.
It means notification will lost the op before ntfimcnd restarted.