lgs: use local mononitic clock instead of getting from active  [#3180]
When NFS is unavailable, client try to write to log. Lgs server
will put it into the queue with the time. This info will check point
with standby. Switch-over happens, and NFS is available again. New active
will get the diffent time b/w current time and the time put into the queue.
The timer on node can be different and it causes the coredump due to
the current time less then time put into the queue.

The fix is to use the local mononitic clock instead of getting from
active.