amfd: process IMM evt higher than MBCSV evt in polling [#3190]
Amfd on standby somehow busy/stuck in a while. The si object was
removed and added again. Then this SI was assigned. Then amfd on standby
continue working again. But it will process event from MBCSV first
(e.g, saAmfSINumCurrActiveAssignments) and it will due to the SI not
created in data based yet.

The solution is to process IMM evt higher priority than MBCSV evt.