ntf: Handle IMM multiple value attribute in notification [#3200]
When IMM changes an attribute with NOTIFY flag, ntfimcn
currently sends the changed attribute values in notification data.
ntfimcn also fetches the old attribute values, corresponding to
the changed attribute values, if exist in IMM, and include the old
attribute values in notification data.

1) With this enahanement, for IMM atrribute with multiple values,
if the number of old attribute values is greater than the number
of new attribute values, the old attribute values will not be
published in the notification data
2) ntftest suite is update