imm: valgrind reports invalid read in imm agent [#2656]
The attrName.size includes the null character in the attrName.buf.
So only 'size' memory should be allocated and copied and not (size+1)

strncpy will copy minimum of attrName.size and attrName.buf, In normal
scenarios this should be attrName.buf

But in case buf is corrupt and extends beyond attrName.size. attrName.size
characters are copied, in this case setting the last character to '\0'
is important