amf: update node member in clm_track_cb [#3257]
During scale-out, node is stopped, 4 events received in order below.
- avd_mds_avnd_up_evh (amfnd up) then clm_track_cb(node join).
At this time node state is still in absent but node member is set to true.
Node member is set in clm_track_cb.
- avd_mds_avnd_down_evh (amfnd down) then clm_track_cb(node left).
The node member is not set to false. So the scale-in failed.

Node member should be set to false in clm_track_cb if amfd receives
node left but node state still in absent.