base: move the testcase sysf_ipc_test to the apitest [#3267]
ncs_leap_startup return error code if the node_id file is missing.
The return error code is expected. But in the ticket-3266 support
for running agent before opensafd. So the test case SysfIpcTest
need to running under opensaf cluster.

In this ticket, move the unit test run as google-test to the apitest