imm: Introduce a configurable attribute saImmSyncrTimeout [#3268]
In a large cluster size, it is impossible to adjust MDS wait time for all
clients in cluster by change IMMA_SYNC_TIMEOUT environment variable.
Because technique environment variable needs to restart to apply the changed
variable and it is a time-consuming.

This ticket introduce a configurable attribute `saImmSyncrTimeout`
to `SaImmMngt` class, that distribute the new timeout to all IMM clients
and force them to use the new timeout. If user deletes value of attribute
`saImmSyncrTimeout`, it will go back to default value (0). And it is will
keep backward compatible with IMMA_SYNC_TIMEOUT technique.
In default, this feature is disabled.

Correct value of variable environment IMMA_SYNCR_TIMEOUT when set a value
less than 10 will only be interpreted as the value 10.