amf: update runtime attributes of node to IMM in sync [#3285]
In large cluster size(lager than 36 nodes), and many
components reside on each node. The runtime of nodes
(AdminState and OperationalState) are take time to
update in IMM and it causes the application get wrong
state of node in IMM instead AMF already update its
data base.

This enhancement is to update these attributes in sync
to IMM.