amfd: Tightens sync window condition to proceed headless restoration [#3271]
In roaming SC cluster, when both active/standby SC go down,
if SC Absence feature is enables, the cluster will nominate
another SC to become active. Amfnd in this nominated SC has
the led_state is true, therefore, amfd in this nominated SC
will proceed the restoration regardless the sync window.

This problem only happens in roaming SC with SC absence, it
does not happen in classical cluster with 2SC-xPLs.

This patch tightens the sync window condition to proceed
the restoration.