log: Introduce the initial clm node status [#3291]
When logd detected loga is up, it will send a callback message to loga.
Loga should wait for the callback message when loga was just created.
This callback helps the messages are processed in right order in logd,
unorder message was mentioned in #1396.
Following is the order of processing messages:
1. logd: agent up message
2. loga: initial clm node status message
3. logd: initial client message
4. logd: final client message
5. logd: agent down message