amf: reboot to recovery PL in split-brain [#3309]
The connection between the standby SC and that PL was dropped,
but that PL still connected with the active SC. It led the
standby SC considered that PL absented regardless the connection
was established after that. During failover, the standby SC will
notify all recorded absent nodes left cluster. It causes PL left
cluster from AMF view but still connect to active.

This scenario is a kind of split-brain use case and amfd should
order PL reboot to recovery the issue.