imm: Fix pbe deadlock [#3314]
Immpbe will fail to restart if a ccb which applies to an immpbe object
changes state to critical. The ccb will be aborted if immpbe restarts
successful. Unfortunately, the pbe couldn't restart if there is an active
ccb in an immpbe object. Solution is to allow immpbe to start even if
there is an active ccb in an immpbe object.