log: delete the empty log file when rotate log file [#3320]
When performing an admin op command with id = 2 (SA_LOG_ADMIN_ROTATE_FILE)
without parameters and there are no log record in the stream.
There will be a empty log file created.

By checking the size of the stream, this ticket avoids creating an empty
log file. Osaflogd will remove the current file If there are no log records
in the stream and create a new file with a new timestamp.