log: Update the destructor and constructor of log agent [#3322]
- The destructor of the static objects will be called in the reverse
order of constructors. If other destructors call a function of the
log agent after the destructor of the log agent, it could be crashed.
Solution is to construct the log agent object early.

- If a process is forked, both processes share all created sockets.
Cleaning resources in one process causes undefined behaviours in
other process. Solution is to empty the destructor of log agent.
Those resources are manually cleaned by the api functions of the
log agent.