ntf: Add new configured variable for logger records timeout [#3363]
Currently, the NTFS_LOGGER_RECORD_TIMEOUT default is set to 10s
and cannot be changed. In some cases, notifications may become overdue
and can be removed.

This commit offer a way for this variable to be configured as an
environment variable that can be exported from a configuration file,
enabling flexible setting.