smf: Fix osafsmfd coredump [#3367]
The node 'inv_id' will be removed in the main thread when the callbacks
are executed successfully. It will then continue to be checked and removed
in the procedure thread if it has not been removed yet. A core dump may occur
if procedures executing in parallel access an invalid pointer to 'inv_id'
after it has already been removed by the main thread.

To avoid this, smfd must avoid removing the node in the sender thread until
the callback has successfully completed and the node has been deleted from
the main thread. The sender thread is only removed when transmitting fails.